03_Implementacao/django-weather/requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
decorator==5.0.9
Django==3.2.5
future==0.18.2
GDAL @ file:///C:/ISEL/6_SemestreVerao2021/PRJ/GDAL-3.3.1-cp38-cp38-win_amd64.whl
geoip2==4.2.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
maxminddb==2.0.3
multidict==5.1.0
numpy==1.21.0
paho-mqtt==1.5.1
pandas==1.3.0
patsy==0.5.1
Pillow==8.3.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
ratelim==0.1.6
requests==2.26.0
scipy==1.7.0
Shapely==1.7.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
statsmodels==0.12.2
tornado==6.1
typing-extensions==3.10.0.0
tzwhere==3.0.3
urllib3==1.26.6
Werkzeug==2.0.1
yarl==1.6.3
